move work to cpp
