+++ src/query/frontend/ast/cypher_main_visitor.hpp
@@ -17,6 +17,7 @@
 #include "query/frontend/opencypher/generated/MemgraphCypherBaseVisitor.h"
 #pragma pop_macro("EOF")  // bring EOF back
 
+#include "query/frontend/ast/ast_storage.hpp"
 #include "query/parameters.hpp"
 #include "utils/exceptions.hpp"
 #include "utils/logging.hpp"
@@ -24,6 +25,13 @@
 #include <string>
 #include <unordered_set>
 #include <utility>
+
+namespace memgraph::query {
+class Query;
+class Expression;
+class Identifier;
+class AuthQuery;
+}  // namespace memgraph::query
 
 namespace memgraph::query::frontend {
 
@@ -39,71 +47,12 @@ class CypherMainVisitor : public antlropencypher::MemgraphCypherBaseVisitor {
       : context_(context), storage_(storage), parameters_(parameters) {}
 
  private:
-  Expression *CreateBinaryOperatorByToken(size_t token, Expression *e1, Expression *e2) {
-    switch (token) {
-      case MemgraphCypher::OR:
-        return storage_->Create<OrOperator>(e1, e2);
-      case MemgraphCypher::XOR:
-        return storage_->Create<XorOperator>(e1, e2);
-      case MemgraphCypher::AND:
-        return storage_->Create<AndOperator>(e1, e2);
-      case MemgraphCypher::PLUS:
-        return storage_->Create<AdditionOperator>(e1, e2);
-      case MemgraphCypher::MINUS:
-        return storage_->Create<SubtractionOperator>(e1, e2);
-      case MemgraphCypher::ASTERISK:
-        return storage_->Create<MultiplicationOperator>(e1, e2);
-      case MemgraphCypher::SLASH:
-        return storage_->Create<DivisionOperator>(e1, e2);
-      case MemgraphCypher::PERCENT:
-        return storage_->Create<ModOperator>(e1, e2);
-      case MemgraphCypher::CARET:
-        return storage_->Create<ExponentiationOperator>(e1, e2);
-      case MemgraphCypher::EQ:
-        return storage_->Create<EqualOperator>(e1, e2);
-      case MemgraphCypher::NEQ1:
-      case MemgraphCypher::NEQ2:
-        return storage_->Create<NotEqualOperator>(e1, e2);
-      case MemgraphCypher::LT:
-        return storage_->Create<LessOperator>(e1, e2);
-      case MemgraphCypher::GT:
-        return storage_->Create<GreaterOperator>(e1, e2);
-      case MemgraphCypher::LTE:
-        return storage_->Create<LessEqualOperator>(e1, e2);
-      case MemgraphCypher::GTE:
-        return storage_->Create<GreaterEqualOperator>(e1, e2);
-      default:
-        throw utils::NotYetImplemented("binary operator");
-    }
-  }
+  Expression *CreateBinaryOperatorByToken(size_t token, Expression *e1, Expression *e2);
 
-  Expression *CreateUnaryOperatorByToken(size_t token, Expression *e) {
-    switch (token) {
-      case MemgraphCypher::NOT:
-        return storage_->Create<NotOperator>(e);
-      case MemgraphCypher::PLUS:
-        return storage_->Create<UnaryPlusOperator>(e);
-      case MemgraphCypher::MINUS:
-        return storage_->Create<UnaryMinusOperator>(e);
-      default:
-        throw utils::NotYetImplemented("unary operator");
-    }
-  }
+  Expression *CreateUnaryOperatorByToken(size_t token, Expression *e);
 
   auto ExtractOperators(std::vector<antlr4::tree::ParseTree *> &all_children,
-                        const std::vector<size_t> &allowed_operators) {
-    std::vector<size_t> operators;
-    for (auto *child : all_children) {
-      antlr4::tree::TerminalNode *operator_node = nullptr;
-      if ((operator_node = dynamic_cast<antlr4::tree::TerminalNode *>(child))) {
-        if (std::find(allowed_operators.begin(), allowed_operators.end(), operator_node->getSymbol()->getType()) !=
-            allowed_operators.end()) {
-          operators.push_back(operator_node->getSymbol()->getType());
-        }
-      }
-    }
-    return operators;
-  }
+                        const std::vector<size_t> &allowed_operators) -> std::vector<size_t>;
 
   /**
    * Convert opencypher's n-ary production to ast binary operators.
