+++ src/storage/v2/indices/edge_type_index.hpp
@@ -14,6 +14,7 @@
 #include "storage/v2/edge_ref.hpp"
 #include "storage/v2/id_types.hpp"
 
+#include <algorithm>
 #include <map>
 #include <vector>
